requirements.txt
praw==7.7.1
anthropic==0.40.0
python-dotenv==1.0.1
requests==2.32.3
